+++ inbm/Changelog.md
@@ -29,6 +29,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - (NEXMANAGE-744) Don't pull in `dmidecode` in `inbm-lib` -- pull in only in `telemetry`
  - (NEXMANAGE-744) Removed all references to `future` library as we do not use Python 2
 
+### Changed
+ - (NEXMANAGE-683) Remove service_type from Operation in common.proto
+
 ## 4.2.5 - 2024-09-04
 ### Fixed
  - Added #!/usr/bin/python3 lines to agents to work in source install mode.